[chore] Cell 네이밍 수정
