+++ Baemin/Baemin/Presentation/Home/List/Model/RankingModel.swift
